SchemaMerger shall only merge into dynamic schemas (#7114)
