+++ common/changes/@itwin/ecschema-editing/schemaediting-merger-only-dynamic-schemas_2024-08-28-09-55.json
@@ -1,0 +1,10 @@
+{
+  "changes": [
+    {
+      "packageName": "@itwin/ecschema-editing",
+      "comment": "",
+      "type": "none"
+    }
+  ],
+  "packageName": "@itwin/ecschema-editing"
+}
