+++ core/ecschema-editing/src/Merging/SchemaMerger.ts
@@ -13,6 +13,7 @@ import { getSchemaDifferences, type SchemaDifferenceResult } from "../Differenci
 import { SchemaMergingVisitor } from "./SchemaMergingVisitor";
 import { SchemaMergingWalker } from "./SchemaMergingWalker";
 import type { SchemaEdits } from "./Edits/SchemaEdits";
+import { ECEditingStatus, SchemaEditingError } from "../Editing/Exception";
 
 /**
  * Defines the context of a Schema merging run.
@@ -79,9 +80,15 @@ export class SchemaMerger {
       );
     }
 
-    const schema = await this._editor.getSchema(targetSchemaKey);
-    if (schema === undefined) {
-      throw new Error(`The target schema '${targetSchemaKey.name}' could not be found in the editing context.`);
+    const schema = await this._editor.getSchema(targetSchemaKey).catch((error: Error) => {
+      if (error instanceof SchemaEditingError && error.errorNumber === ECEditingStatus.SchemaNotFound) {
+        throw new Error(`The target schema '${targetSchemaKey.name}' could not be found in the editing context.`);
+      }
+      throw error;
+    });
+
+    if (!schema.customAttributes || !schema.customAttributes.has("CoreCustomAttributes.DynamicSchema")) {
+      throw new Error(`The target schema '${targetSchemaKey.name}' is not dynamic. Only dynamic schemas are supported for merging.`);
     }
 
     const visitor = new SchemaMergingVisitor({
