+++ core/ecschema-editing/src/test/Merging/UnitSystemMerger.test.ts
@@ -2,18 +2,25 @@
 * Copyright (c) Bentley Systems, Incorporated. All rights reserved.
 * See LICENSE.md in the project root for license terms and full copyright notice.
 *--------------------------------------------------------------------------------------------*/
-import { Schema, SchemaContext, SchemaItemType, UnitSystem } from "@itwin/ecschema-metadata";
+import { Schema, SchemaItemType, UnitSystem } from "@itwin/ecschema-metadata";
 import { SchemaMerger } from "../../Merging/SchemaMerger";
 import { expect } from "chai";
+import { BisTestHelper } from "../TestUtils/BisTestHelper";
 
 describe("Unit system merger tests", () => {
   it("should merge missing unit system", async () => {
-    const targetContext = new SchemaContext();
+    const targetContext = await BisTestHelper.getNewContext();
     await Schema.fromJson({
       $schema: "https://dev.bentley.com/json_schemas/ec/32/ecschema",
       name: "TargetSchema",
       version: "1.0.0",
       alias: "target",
+      references: [
+        { name: "CoreCustomAttributes", version: "01.00.01" },
+      ],
+      customAttributes: [
+        { className: "CoreCustomAttributes.DynamicSchema" },
+      ],
     }, targetContext);
 
     const merger = new SchemaMerger(targetContext);
@@ -31,19 +38,26 @@ describe("Unit system merger tests", () => {
       }],
     });
 
-    const mergedUnitSystem = await mergedSchema.getItem("testUnitSystem") as UnitSystem;
-    expect(mergedUnitSystem).is.not.undefined;
-    expect(mergedUnitSystem.label).equals("Imperial");
-    expect(mergedUnitSystem.description).equals("Imperial Unit System");
+    await expect(mergedSchema.getItem("testUnitSystem")).to.be.eventually.not.undefined
+      .then((mergedUnitSystem: UnitSystem) => {
+        expect(mergedUnitSystem).to.have.a.property("label", "Imperial");
+        expect(mergedUnitSystem).to.have.a.property("description", "Imperial Unit System");
+      });
   });
 
   it("should merge unit system with new label and description", async () => {
-    const targetContext = new SchemaContext();
+    const targetContext = await BisTestHelper.getNewContext();
     await Schema.fromJson({
       $schema: "https://dev.bentley.com/json_schemas/ec/32/ecschema",
       name: "TargetSchema",
       version: "1.0.0",
       alias: "target",
+      references: [
+        { name: "CoreCustomAttributes", version: "01.00.01" },
+      ],
+      customAttributes: [
+        { className: "CoreCustomAttributes.DynamicSchema" },
+      ],
       items: {
         testUnitSystem: {
           schemaItemType: "UnitSystem",
@@ -66,9 +80,10 @@ describe("Unit system merger tests", () => {
       }],
     });
 
-    const mergedUnitSystem = await mergedSchema.getItem("testUnitSystem") as UnitSystem;
-    expect(mergedUnitSystem).is.not.undefined;
-    expect(mergedUnitSystem.label).equals("New Imperial");
-    expect(mergedUnitSystem.description).equals("New Imperial Unit System");
+    await expect(mergedSchema.getItem("testUnitSystem")).to.be.eventually.not.undefined
+      .then((mergedUnitSystem: UnitSystem) => {
+        expect(mergedUnitSystem).to.have.a.property("label", "New Imperial");
+        expect(mergedUnitSystem).to.have.a.property("description", "New Imperial Unit System");
+      });
   });
 });
