+++ core/ecschema-editing/src/test/TestUtils/BisTestHelper.ts
@@ -30,6 +30,13 @@ const coreCustomAttributesSchema = {
       modifier: "sealed",
       schemaItemType: "CustomAttributeClass",
     },
+    DynamicSchema: {
+      schemaItemType: "CustomAttributeClass",
+      label: "Dynamic Schema",
+      description: "Identifies a schema as dynamically generated by an application.  Like versions of a schema may differ and are not guaranteed to merge without conflicts.",
+      modifier: "Sealed",
+      appliesTo: "Schema",
+    },
   },
 };
